+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cache;
 mod fun;
+pub mod intranet_nets;
 pub mod mem;
 mod redis;
 pub use crate::mem::DBQuery;